// src/features/home/Hero/HeroVariants.d.ts
export declare const HeroVariants: (props?: ({
    size?: "sm" | "md" | "lg" | null | undefined;
    alignment?: "center" | "left" | null | undefined;
    imagePosition?: "left" | "right" | null | undefined;
} & import("class-variance-authority/types").ClassProp) | undefined) => string;
export declare const heroGridVariants: (props?: import("class-variance-authority/types").ClassProp | undefined) => string;
export declare const heroContentVariants: (props?: ({
    alignment?: "center" | "left" | null | undefined;
} & import("class-variance-authority/types").ClassProp) | undefined) => string;
export declare const heroBadgeVariants: (props?: import("class-variance-authority/types").ClassProp | undefined) => string;
export declare const heroTitleVariants: (props?: import("class-variance-authority/types").ClassProp | undefined) => string;
export declare const heroSubtitleVariants: (props?: import("class-variance-authority/types").ClassProp | undefined) => string;
export declare const heroDescriptionVariants: (props?: import("class-variance-authority/types").ClassProp | undefined) => string;
export declare const heroImageVariants: (props?: import("class-variance-authority/types").ClassProp | undefined) => string;
export declare const heroImageElementVariants: (props?: import("class-variance-authority/types").ClassProp | undefined) => string;
export declare const heroBackgroundVariants: (props?: import("class-variance-authority/types").ClassProp | undefined) => string;
export declare const heroCircleVariants: (props?: ({
    position?: "topLarge" | "topSmall" | "bottomLarge" | null | undefined;
} & import("class-variance-authority/types").ClassProp) | undefined) => string;
export declare const heroCarouselVariants: (props?: import("class-variance-authority/types").ClassProp | undefined) => string;
export declare const heroViewportVariants: (props?: import("class-variance-authority/types").ClassProp | undefined) => string;
export declare const heroTrackVariants: (props?: import("class-variance-authority/types").ClassProp | undefined) => string;
export declare const heroSlideVariants: (props?: import("class-variance-authority/types").ClassProp | undefined) => string;
export declare const heroArrowContainerVariants: (props?: ({
    side?: "left" | "right" | null | undefined;
    visibility?: "always" | "desktop" | null | undefined;
} & import("class-variance-authority/types").ClassProp) | undefined) => string;
export declare const heroArrowButtonVariants: (props?: ({
    size?: "sm" | "md" | "lg" | null | undefined;
} & import("class-variance-authority/types").ClassProp) | undefined) => string;
export declare const heroPaginationVariants: (props?: import("class-variance-authority/types").ClassProp | undefined) => string;
export declare const heroPaginationDotVariants: (props?: ({
    active?: boolean | null | undefined;
    size?: "sm" | "md" | "lg" | null | undefined;
} & import("class-variance-authority/types").ClassProp) | undefined) => string;